Rename FileSystemJobState to FileSystemGraphObjectStore
